me cago en long

+++ src/main/java/com/proyecto/demo/repository/BarcoRepository.java
@@ -1,4 +1,4 @@
 @Repository
-public interface BarcoRepository extends JpaRepository<Barco, Integer> {
+public interface BarcoRepository extends JpaRepository<Barco, Long> {
     // Aquí puedes definir métodos personalizados si es necesario
 }
